Simplify usage via builder harness
